Building fails at NUnitLite when using the Visual Studio 2013 x64 command prompt. Looking at the failure, it is because /define:TRACE;DEBUG;NET_4_5;NUNITLITE is not included by msbuild in the csc.exe command line.
This is because vcvarsall.bat sets the environment variable Platform=X64, but doesn't set it for the X86 command prompt. I assume that msbuild then interprets this as the Platform to build which doesn't exist in our project files, thus no defines.

I fixed this by just adding Platform=AnyCpu to the msbuild properties to force it.