Display warnings even when there are only warnings
The "Errors, Failures and Warnings" section is now also written,
if there are tests with warnings, but no tests failed.

Note that we now only create ConsoleTestResult if the site is
different from `null` as warnings do not set the site.
Otherwise we will report warnings "One or more child tests had warnings"
for "test-suite"-elements above tests with warnings.

Fixes #289