Fix paths in addins files for global nuget package location
Paths on Linux are case-sensitive and nuget converts the paths
to lowercase when installing them in the global package location.
However, when installing the packages in a local directory the
paths don't get changed.

This change adds the lowercase paths in addition to the camel case
ones so that we find extensions that are installed parallel to
the NUnit.ConsoleRunner nuget package regardless of whether it is
installed in a local directory or in the global package location.

This fixes #775.